sudokuTS-15 Implement XYZ-Wing rule: Moved tough abstract rules.
Moved tough abstract rules to its own class.
